@@ -20,35 +20,37 @@
     "graphql"
   ],
   "peerDependencies": {
-    "@nestjs/common": "^7.0.0",
-    "@nestjs/core": "^7.0.0",
-    "@nestjs/graphql": "^7.0.0",
-    "graphql": "^14.1.1",
+    "@apollo/gateway": "^0.41.0",
+    "@nestjs/common": "^8.0.6",
+    "@nestjs/core": "^8.0.6",
+    "@nestjs/graphql": "^9.0.4",
+    "graphql": "^15.5.3",
     "reflect-metadata": "^0.1.12"
   },
   "dependencies": {
-    "dataloader": "^2.0.0",
-    "rxjs": "^6.5.4"
+    "dataloader": "^2.0.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "^7.1.2",
-    "@nestjs/common": "^7.0.5",
-    "@nestjs/core": "^7.0.5",
-    "@nestjs/graphql": "^7.0.15",
-    "@nestjs/platform-express": "^7.0.7",
-    "@nestjs/testing": "^7.0.7",
-    "@nestjs/typeorm": "^7.0.0",
-    "@types/jest": "^25.2.1",
-    "apollo-server-express": "^2.9.16",
-    "apollo-server-testing": "^2.11.0",
-    "graphql": "^14.1.1",
-    "jest": "^25.2.7",
+    "@apollo/gateway": "^0.41.0",
+    "@nestjs/cli": "^8.1.1",
+    "@nestjs/common": "^8.0.6",
+    "@nestjs/core": "^8.0.6",
+    "@nestjs/graphql": "^9.0.4",
+    "@nestjs/platform-express": "^8.0.6",
+    "@nestjs/testing": "^8.0.6",
+    "@nestjs/typeorm": "^8.0.2",
+    "@types/jest": "^27.0.1",
+    "apollo-server-express": "^3.3.0",
+    "graphql": "^15.5.3",
+    "jest": "^27.2.0",
     "reflect-metadata": "^0.1.12",
-    "sqlite3": "^4.1.1",
-    "ts-jest": "^25.3.1",
+    "rxjs": "^7.3.0",
+    "sqlite3": "^5.0.2",
+    "ts-jest": "^27.0.5",
+    "ts-morph": "^12.0.0",
     "typeorm": "^0.2.24",
     "typeorm-factory": "^0.0.14",
-    "typescript": "^3.7.4"
+    "typescript": "^4.4.3"
   },
   "types": "index.d.ts",
   "jest": {
